backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
python-multipart
httpx
